# specs/001-homelab-dashboard/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: HomeDash API
  version: 0.1.0
  description: |
    API contracts for HomeDash.

    Notes:
    - Cookie-based sessions are used for authentication.
    - CSRF protection is required for state-changing requests.
servers:
  - url: http://localhost:3000
...
paths:
...
  /api/user/preferences:
    get:
      tags: [user]
      summary: Get user preferences
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
        '401':
          description: Not authenticated
    put:
      tags: [user]
      summary: Update user preferences
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPreferencesUpdateRequest'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPreferences'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: homedash_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
  schemas:
...
    UserPreferences:
      type: object
      required: [themeMode]
      properties:
        themeMode:
          type: string
          enum: [light, dark]
        webDashboardId:
          type: string
          nullable: true
        mobileDashboardId:
          type: string
          nullable: true
    UserPreferencesUpdateRequest:
      type: object
      properties:
        themeMode:
          type: string
          enum: [light, dark]
        webDashboardId:
          type: string
          nullable: true
        mobileDashboardId:
          type: string
          nullable: true
...